Move User Interface classes to folder ui. Modify imports
