+++ ui/giscube_admin_configure_dialog_base.ui
