+++ ui/giscube_admin_dockwidget.py
