+++ ui/giscube_admin_dockwidget_base.ui
