+++ ui/new_server_dialog_base.ui
